refactor: 사용자의 게시글 목록 조회 기능 리팩토링 (#297)
